+++ src/app/gatherings/[id]/page.tsx
@@ -1,3 +1,3 @@
-import GatheringsDetail from '@/pages/gatherings-detail';
-
-export default GatheringsDetail;
+export default function GatheringsDetail() {
+  return <div>모임 상세 페이지</div>;
+}
