+++ src/app/gatherings/page.tsx
@@ -1,3 +1,3 @@
-import Gatherings from '@/pages/gatherings';
-
-export default Gatherings;
+export default function Gatherings() {
+  return <div>모임 메인 페이지</div>;
+}
