+++ src/app/signin/page.tsx
@@ -1,3 +1,3 @@
-import Signin from '@/pages/signin';
-
-export default Signin;
+export default function Signin() {
+  return <div>로그인 페이지</div>;
+}
