+++ src/app/signup/page.tsx
@@ -1,3 +1,3 @@
-import Signup from '@/pages/signup';
-
-export default Signup;
+export default function Signup() {
+  return <div>회원가입 페이지</div>;
+}
