--- src/pages/_error.jsx
@@ -1,17 +1,0 @@
-import * as Sentry from "@sentry/nextjs";
-import Error from "next/error";
-
-const CustomErrorComponent = (props) => {
-  return <Error statusCode={props.statusCode} />;
-};
-
-CustomErrorComponent.getInitialProps = async (contextData) => {
-  // In case this is running in a serverless function, await this in order to give Sentry
-  // time to send the error before the lambda exits
-  await Sentry.captureUnderscoreErrorException(contextData);
-
-  // This will contain the status code of the response
-  return Error.getInitialProps(contextData);
-};
-
-export default CustomErrorComponent;
