--- src/pages/gatherings-detail.tsx
@@ -1,3 +1,0 @@
-export default function GatheringsDetail() {
-  return <div>모임 상세 페이지</div>;
-}
