--- src/pages/gatherings.tsx
@@ -1,3 +1,0 @@
-export default function Gatherings() {
-  return <div>모임 메인 페이지</div>;
-}
