--- src/pages/like-gatherings.tsx
@@ -1,3 +1,0 @@
-export default function LikeGatherings() {
-  return <div>찜한 모임 페이지</div>;
-}
