--- src/pages/mypage.tsx
@@ -1,3 +1,0 @@
-export default function Mypage() {
-  return <div>마이페이지</div>;
-}
