--- src/pages/reviews.tsx
@@ -1,3 +1,0 @@
-export default function Reviews() {
-  return <div>리뷰 페이지</div>;
-}
