--- src/pages/signin.tsx
@@ -1,3 +1,0 @@
-export default function Signin() {
-  return <div>로그인 페이지</div>;
-}
