--- src/pages/signup.tsx
@@ -1,3 +1,0 @@
-export default function Signup() {
-  return <div>회원가입 페이지</div>;
-}
